# Conflicts:
#	src/main/java/com/softserve/maklertaboo/constant/ErrorMessage.java
#	src/main/java/com/softserve/maklertaboo/service/FlatService.java

+++ src/main/java/com/softserve/maklertaboo/controller/FlatController.java
@@ -44,7 +44,7 @@ public class FlatController {
     }
 
     @GetMapping("{postId}")
-    public FlatDetailDto getActive(@PathVariable Integer postId) {
+    public FlatDetailDto getActive(@PathVariable Long postId) {
         return flatDetailMapper.convertToDto(flatService.getById(postId));
     }
